[wordSwap-6] - Retirada de compentes que usam 'use client' do layout
