--- pdf-worker.js
@@ -1,1 +1,0 @@
-export const WORKERSRC = 'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.16.105/pdf.worker.min.js'
