+++ src/components/Modal/UpdateName.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as React from 'react';
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
