+++ src/components/SideBarFillTemplate.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Button,
   Drawer,
