+++ src/components/patterns/components/ThemeAndCssProvider.tsx
@@ -1,0 +1,16 @@
+import { ThemeProvider } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
+import { darkTheme } from '@/theme';
+
+export const ThemeAndCssProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  return (
+    <ThemeProvider theme={darkTheme}>
+      <CssBaseline />
+      {children}
+    </ThemeProvider>
+  );
+};
