+++ src/service/axiosApi.ts
@@ -2,7 +2,7 @@ import axios from "axios"
 
 
 const api = axios.create({
-  baseURL: "http://localhost:3000/",
+  baseURL: process.env.NEXT_PUBLIC_URL,
 });
 
 export default api;
